refactor sorters into sortHelper
