+++ test/unit/helper/gameHelper.ts
@@ -4,19 +4,15 @@ import { render, screen } from "@testing-library/react";
 import { JapaneseText } from "../../../src/helper/JapaneseText";
 import {
   activeGroupIncludes,
-  alphaOrder,
-  difficultySubFilter,
   getJapaneseHint,
   getVerbFormsArray,
-  randomOrder,
-  spaceRepOrder,
   termFilterByType,
 } from "../../../src/helper/gameHelper";
 import { TermFilterBy } from "../../../src/slices/settingHelper";
-import type { RawVocabulary } from "../../../src/typings/raw";
+import type { Optional, RawVocabulary } from "../../../src/typings/raw";
 
 describe("gameHelper", function () {
-  const terms: RawVocabulary[] = [
+  const terms: Optional<RawVocabulary,'tags'>[] = [
     {
       english: "blue",
       grp: "Noun",
@@ -122,470 +118,6 @@ describe("gameHelper", function () {
       uid: "ee5b790c89a6e8811e7b3c97ee79534c",
     },
   ];
-  describe("spaceRepOrder", function () {
-    it("term order when undefined", function () {
-      const expected = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0];
-      const spaceRepObj = {
-        "00c102a7e10b45b19afbab71c030bf63": {
-          vC: 1,
-          lastView: "2021-09-28T17:38:09.319Z",
-        },
-      };
-
-      const actual = spaceRepOrder(terms, spaceRepObj);
-      expect(actual).to.deep.eq(expected);
-    });
-    it("date order oldest first", function () {
-      const expected = [12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0];
-
-      const spaceRepObj = {
-        "00c102a7e10b45b19afbab71c030bf63": {
-          // english: 'blue',
-          vC: 1,
-          lastView: "2020-01-01T01:12:00.000Z",
-        },
-        "12e960a0d8ae82cf5804a8e9f192d664": {
-          // english: 'purple',
-          vC: 1,
-          lastView: "2020-01-01T01:11:00.000Z",
-        },
-        "3caebfbcc0613501e39f7819e9b2d5a3": {
-          // english: 'yellow',
-          vC: 1,
-          lastView: "2020-01-01T01:10:00.000Z",
-        },
-        "3e30afabc4b11c821e398d6ee1226c59": {
-          // english: 'white',
-          vC: 1,
-          lastView: "2020-01-01T01:09:00.000Z",
-        },
-        "6fca55dd4a82b78256cc9c22e2934938": {
-          // english: 'green',
-          vC: 1,
-          lastView: "2020-01-01T01:08:00.000Z",
-        },
-        "729307b04a77bccc5db86d6b49f55f2f": {
-          // english: 'grey',
-          vC: 1,
-          lastView: "2020-01-01T01:07:00.000Z",
-        },
-        "792bc286417b7ac5cfee4bdd7872f892": {
-          // english: 'beige',
-          vC: 1,
-          lastView: "2020-01-01T01:06:00.000Z",
-        },
-        a8787c646e16617ba878fe462e4d1ffe: {
-          // english: 'black',
-          vC: 1,
-          lastView: "2020-01-01T01:05:00.000Z",
-        },
-        b78d1a33132a9090a7b545f7aa3d2f63: {
-          // english: 'brown',
-          vC: 1,
-          lastView: "2020-01-01T01:04:00.000Z",
-        },
-        de108914e6e86883ed97dc62918ae89a: {
-          // english: 'grey',
-          vC: 1,
-          lastView: "2020-01-01T01:03:00.000Z",
-        },
-        e5d47019e1b948c2445b6c1ea3850c2b: {
-          // english: 'red',
-          vC: 1,
-          lastView: "2020-01-01T01:02:00.000Z",
-        },
-        e86638b52f2028b1ff3685e13bfd71ac: {
-          // english: 'orange',
-          vC: 1,
-          lastView: "2020-01-01T01:01:00.000Z",
-        },
-        ee5b790c89a6e8811e7b3c97ee79534c: {
-          // english: 'pink',
-          vC: 1,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-      };
-
-      const actual = spaceRepOrder(terms, spaceRepObj);
-      expect(actual).to.deep.eq(expected);
-    });
-    it("view count order lowest first", function () {
-      const expected = [12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0];
-      const spaceRepObj = {
-        "00c102a7e10b45b19afbab71c030bf63": {
-          // english: 'blue',
-          vC: 13,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        "12e960a0d8ae82cf5804a8e9f192d664": {
-          // english: 'purple',
-          vC: 12,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        "3caebfbcc0613501e39f7819e9b2d5a3": {
-          // english: 'yellow',
-          vC: 11,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        "3e30afabc4b11c821e398d6ee1226c59": {
-          // english: 'white',
-          vC: 10,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        "6fca55dd4a82b78256cc9c22e2934938": {
-          // english: 'green',
-          vC: 9,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        "729307b04a77bccc5db86d6b49f55f2f": {
-          // english: 'grey',
-          vC: 8,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        "792bc286417b7ac5cfee4bdd7872f892": {
-          // english: 'beige',
-          vC: 7,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        a8787c646e16617ba878fe462e4d1ffe: {
-          // english: 'black',
-          vC: 6,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        b78d1a33132a9090a7b545f7aa3d2f63: {
-          // english: 'brown',
-          vC: 5,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        de108914e6e86883ed97dc62918ae89a: {
-          // english: 'grey',
-          vC: 4,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        e5d47019e1b948c2445b6c1ea3850c2b: {
-          // english: 'red',
-          vC: 3,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        e86638b52f2028b1ff3685e13bfd71ac: {
-          // english: 'orange',
-          vC: 2,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        ee5b790c89a6e8811e7b3c97ee79534c: {
-          // english: 'pink',
-          vC: 1,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-      };
-
-      const actual = spaceRepOrder(terms, spaceRepObj);
-      expect(actual).to.deep.eq(expected);
-    });
-    it("date order superseedes count order", function () {
-      const expected = [11, 12, 10, 9, 8, 7, 6, 5, 4, 3, 0, 1, 2];
-      const spaceRepObj = {
-        "00c102a7e10b45b19afbab71c030bf63": {
-          // english: 'blue',
-          vC: 1,
-          lastView: "2020-01-01T01:10:00.000Z",
-        },
-        "12e960a0d8ae82cf5804a8e9f192d664": {
-          // english: 'purple',
-          vC: 2,
-          lastView: "2020-01-01T01:10:00.000Z",
-        },
-        "3caebfbcc0613501e39f7819e9b2d5a3": {
-          // english: 'yellow',
-          vC: 2,
-          lastView: "2020-01-01T01:10:00.000Z",
-        },
-        "3e30afabc4b11c821e398d6ee1226c59": {
-          // english: 'white',
-          vC: 1,
-          lastView: "2020-01-01T01:09:00.000Z",
-        },
-        "6fca55dd4a82b78256cc9c22e2934938": {
-          // english: 'green',
-          vC: 1,
-          lastView: "2020-01-01T01:08:00.000Z",
-        },
-        "729307b04a77bccc5db86d6b49f55f2f": {
-          // english: 'grey',
-          vC: 1,
-          lastView: "2020-01-01T01:07:00.000Z",
-        },
-        "792bc286417b7ac5cfee4bdd7872f892": {
-          // english: 'beige',
-          vC: 1,
-          lastView: "2020-01-01T01:06:00.000Z",
-        },
-        a8787c646e16617ba878fe462e4d1ffe: {
-          // english: 'black',
-          vC: 1,
-          lastView: "2020-01-01T01:05:00.000Z",
-        },
-        b78d1a33132a9090a7b545f7aa3d2f63: {
-          // english: 'brown',
-          vC: 1,
-          lastView: "2020-01-01T01:04:00.000Z",
-        },
-        de108914e6e86883ed97dc62918ae89a: {
-          // english: 'grey',
-          vC: 1,
-          lastView: "2020-01-01T01:03:00.000Z",
-        },
-        e5d47019e1b948c2445b6c1ea3850c2b: {
-          // english: 'red',
-          vC: 1,
-          lastView: "2020-01-01T01:02:00.000Z",
-        },
-        e86638b52f2028b1ff3685e13bfd71ac: {
-          // english: 'orange',
-          vC: 1,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        ee5b790c89a6e8811e7b3c97ee79534c: {
-          // english: 'pink',
-          vC: 1,
-          lastView: "2020-01-01T01:01:00.000Z",
-        },
-      };
-
-      const actual = spaceRepOrder(terms, spaceRepObj);
-      expect(actual).to.deep.eq(expected);
-    });
-    it("timed play ", function () {
-      const expected = [0, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 12];
-      const spaceRepObj = {
-        "00c102a7e10b45b19afbab71c030bf63": {
-          // english: 'blue',
-          vC: 13,
-          lastView: "2020-01-01T01:00:00.000Z",
-          tpAcc: 0.649, // below .65 sorted to the front
-        },
-        "12e960a0d8ae82cf5804a8e9f192d664": {
-          // english: 'purple',
-          vC: 12,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        "3caebfbcc0613501e39f7819e9b2d5a3": {
-          // english: 'yellow',
-          vC: 11,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        "3e30afabc4b11c821e398d6ee1226c59": {
-          // english: 'white',
-          vC: 10,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        "6fca55dd4a82b78256cc9c22e2934938": {
-          // english: 'green',
-          vC: 9,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        "729307b04a77bccc5db86d6b49f55f2f": {
-          // english: 'grey',
-          vC: 8,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        "792bc286417b7ac5cfee4bdd7872f892": {
-          // english: 'beige',
-          vC: 7,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        a8787c646e16617ba878fe462e4d1ffe: {
-          // english: 'black',
-          vC: 6,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        b78d1a33132a9090a7b545f7aa3d2f63: {
-          // english: 'brown',
-          vC: 5,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        de108914e6e86883ed97dc62918ae89a: {
-          // english: 'grey',
-          vC: 4,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        e5d47019e1b948c2445b6c1ea3850c2b: {
-          // english: 'red',
-          vC: 3,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        e86638b52f2028b1ff3685e13bfd71ac: {
-          // english: 'orange',
-          vC: 2,
-          lastView: "2020-01-01T01:00:00.000Z",
-        },
-        ee5b790c89a6e8811e7b3c97ee79534c: {
-          // english: 'pink',
-          vC: 1,
-          lastView: "2020-01-01T01:00:00.000Z",
-          tpAcc: 0.65, // .65 goes to the end
-        },
-      };
-
-      const actual = spaceRepOrder(terms, spaceRepObj);
-      expect(actual).to.deep.eq(expected);
-    });
-    it("has right type", function () {
-      const wrongType = [
-        "1",
-        "2",
-        "3",
-        "4",
-        "5",
-        "6",
-        "7",
-        "8",
-        "9",
-        "10",
-        "11",
-        "12",
-        "0",
-      ];
-      const spaceRepObj = {
-        "00c102a7e10b45b19afbab71c030bf63": {
-          vC: 1,
-          lastView: "2021-09-28T17:38:09.319Z",
-        },
-      };
-
-      const actual = spaceRepOrder(terms, spaceRepObj);
-      expect(actual).to.not.deep.eq(wrongType);
-    });
-  });
-  describe("alphaOrder", function () {
-    it("has right order", function () {
-      const expected = [0, 10, 2, 7, 3, 8, 5, 9, 4, 1, 11, 12, 6];
-      const expectedJ = [
-        "00c102a7e10b45b19afbab71c030bf63",
-        "e5d47019e1b948c2445b6c1ea3850c2b",
-        "3caebfbcc0613501e39f7819e9b2d5a3",
-        "a8787c646e16617ba878fe462e4d1ffe",
-        "3e30afabc4b11c821e398d6ee1226c59",
-        "b78d1a33132a9090a7b545f7aa3d2f63",
-        "729307b04a77bccc5db86d6b49f55f2f",
-        "de108914e6e86883ed97dc62918ae89a",
-        "6fca55dd4a82b78256cc9c22e2934938",
-        "12e960a0d8ae82cf5804a8e9f192d664",
-        "e86638b52f2028b1ff3685e13bfd71ac",
-        "ee5b790c89a6e8811e7b3c97ee79534c",
-        "792bc286417b7ac5cfee4bdd7872f892",
-      ];
-      const expectedE = [
-        "792bc286417b7ac5cfee4bdd7872f892",
-        "a8787c646e16617ba878fe462e4d1ffe",
-        "00c102a7e10b45b19afbab71c030bf63",
-        "b78d1a33132a9090a7b545f7aa3d2f63",
-        "6fca55dd4a82b78256cc9c22e2934938",
-        "729307b04a77bccc5db86d6b49f55f2f",
-        "de108914e6e86883ed97dc62918ae89a",
-        "e86638b52f2028b1ff3685e13bfd71ac",
-        "ee5b790c89a6e8811e7b3c97ee79534c",
-        "12e960a0d8ae82cf5804a8e9f192d664",
-        "e5d47019e1b948c2445b6c1ea3850c2b",
-        "3e30afabc4b11c821e398d6ee1226c59",
-        "3caebfbcc0613501e39f7819e9b2d5a3",
-      ];
-
-      const { order, jOrder, eOrder } = alphaOrder(terms);
-      expect(order, "order").to.deep.eq(expected);
-
-      const jOrderUids = jOrder.map((j) => j.uid);
-      expect(jOrderUids, "Japanse order").to.deep.eq(expectedJ);
-
-      const eOrderUids = eOrder.map((e) => e.uid);
-      expect(eOrderUids, "English order").to.deep.eq(expectedE);
-    });
-  });
-  describe("randomOrder", function () {
-    it("should not be ordered", function () {
-      const expected = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12];
-
-      const actual = randomOrder(terms);
-      expect(actual).to.not.deep.eq(expected); //should not be ordered
-      expect(actual).to.include.members(expected); //should contain all originals
-    });
-  });
-  describe("difficultySubFilter", function () {
-    const terms = [{ uid: "a" }, { uid: "b" }, { uid: "c" }, { uid: "d" }];
-
-    it("below a value", function () {
-      const expected = [{ uid: "a" }];
-      const tMeta = {
-        a: { difficulty: 36 },
-        b: { difficulty: 90 },
-        c: { difficulty: 90 },
-        d: { difficulty: 90 },
-      };
-
-      const actual = difficultySubFilter(-50, terms, tMeta);
-      expect(actual).to.deep.eq(expected);
-    });
-    it("undefined is below", function () {
-      const expected = [{ uid: "a" }];
-      const tMeta = {
-        a: { difficulty: undefined },
-        b: { difficulty: 90 },
-        c: { difficulty: 90 },
-        d: { difficulty: 90 },
-      };
-
-      const actual = difficultySubFilter(-80, terms, tMeta);
-      expect(actual).to.deep.eq(expected);
-    });
-    it("above a value", function () {
-      const expected = [{ uid: "b" }, { uid: "c" }, { uid: "d" }];
-      const tMeta = {
-        a: { difficulty: 36 },
-        b: { difficulty: 90 },
-        c: { difficulty: 90 },
-        d: { difficulty: 90 },
-      };
-
-      const actual = difficultySubFilter(50, terms, tMeta);
-      expect(actual).to.deep.eq(expected);
-    });
-    it("undefined is above", function () {
-      const expected = [{ uid: "b" }, { uid: "c" }, { uid: "d" }];
-      const tMeta = {
-        a: { difficulty: 6 },
-        b: { difficulty: undefined },
-        c: { difficulty: undefined },
-        d: { difficulty: undefined },
-      };
-
-      const actual = difficultySubFilter(25, terms, tMeta);
-      expect(actual).to.deep.eq(expected);
-    });
-    it("mixed", function () {
-      const expected = [{ uid: "a" }, { uid: "b" }, { uid: "c" }];
-      const tMeta = {
-        a: { difficulty: 6 },
-        b: { difficulty: undefined },
-        c: { difficulty: 31 },
-        d: { difficulty: 80 },
-      };
-
-      const actual = difficultySubFilter(-75, terms, tMeta);
-      expect(actual).to.deep.eq(expected);
-    });
-    it("empty result", function () {
-      const expected = [];
-      const tMeta = {
-        a: { difficulty: 36 },
-        b: { difficulty: 90 },
-        c: { difficulty: 90 },
-        d: { difficulty: 90 },
-      };
-
-      const actual = difficultySubFilter(90, terms, tMeta);
-      expect(actual).to.deep.eq(expected);
-    });
-  });
   describe("termFilterByType", function () {
     const timeGrpTerms = [
       {
